% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psbt.R
\name{psbt}
\alias{psbt}
\alias{psbt1}
\alias{psbt2}
\alias{psbt3}
\alias{psbt4}
\title{Owen distribution functions}
\usage{
psbt1(nu, t1, t2, delta1, delta2)

psbt2(nu, t1, t2, delta1, delta2)

psbt3(nu, t1, t2, delta1, delta2)

psbt4(nu, t1, t2, delta1, delta2)
}
\arguments{
\item{nu}{integer greater than \eqn{1}, the number of degrees of freedom;
infinite allowed}

\item{t1, t2}{two finite numbers, positive or negative}

\item{delta1, delta2}{two vectors of possibly infinite numbers with the same length,
the noncentrality parameters}
}
\value{
A vector of numbers between \eqn{0} and \eqn{1}.
}
\description{
Evaluates the Owen cumulative distribution function
for an integer number of degrees of freedom.
\itemize{
\item \code{psbt1} evaluates \eqn{P(T_1\le t_1, T_2 \le t_2)}
\item \code{psbt2} evaluates \eqn{P(T_1\le t_1, T_2 \ge t_2)}
\item \code{psbt3} evaluates \eqn{P(T_1\ge t_1, T_2 \ge t_2)}
\item \code{psbt4} evaluates \eqn{P(T_1\ge t_1, T_2 \le t_2)}
}
}
\note{
When the number of degrees of freedom is odd, the procedure resorts to
the Owen T-function (\code{\link{OwenT}}).
}
\examples{
nu=5; t1=1; t2=2; delta1=2; delta2=3
( p1 <- psbt1(nu, t1, t2, delta1, delta2) )
( p2 <- psbt2(nu, t1, t2, delta1, delta2) )
( p3 <- psbt3(nu, t1, t2, delta1, delta2) )
( p4 <- psbt4(nu, t1, t2, delta1, delta2) )
# the sum should be 1
p1+p2+p3+p4
}
\references{
Owen, D. B. (1965).
A special case of a bivariate noncentral t-distribution.
\emph{Biometrika} \bold{52}, 437-446.
}
\seealso{
It is better to use \code{\link{powen}} if \code{delta1>delta2}.
}
